Generate member array initializers
Initialization of member variables with specified number of elements is
done in a special method named "2" if the class has a constructor, or in
a generated constructor otherwise.

The call to method "2" is inserted at the beginning of the user-defined
constructor. This is done regardless of whether the class has array
members or not.