Allow anonymous structs in HLL declarations
Since the HLL0 section of ain (< v14) does not have information that
specifies struct type, AinDecompiler generates a bare struct keyword as
a type specifier in .hll files.

To simplify the parser, this implementation uses a hack to replace
`struct` keywords in hll with HLL_STRUCT.

In the type checker, hll_struct matches any struct.