[Decompiler] Fix out-of-function jumps in MG Rance 6
MangaGamer Rance 6 has two SP_CREATE_REAL and SP_SET_CG_REAL functions,
and the new functions have jump instructions into the old functions.

The new and old functions are identical except for the addresses, so
we'll fix this by replacing the target addresses of these jump
instructions with those of the new functions.