[Decompiler] Link lambda functions via AST scan
The previous method of determining parent-child relationships for lambda
functions by their position in the code section was unreliable. For
instance, a lambda for activity::detail::CUserComponentActivity@2 in
Rance10.ain was defined far from its parent.

This changes the decompiler to scan a function's decompiled AST to
identify and link contained lambda functions.